Разделил код на отдельные модули.
